requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
requests==2.32.3
streamlit==1.40.0
pillow==11.0.0
